requirements.txt
asttokens==2.0.5
certifi==2021.10.8
click==8.1.2
colorama==0.4.4
elasticsearch==7.16.3
executing==0.8.3
Flask==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
pure-eval==0.2.2
pydantic==1.9.0
python-dotenv==0.20.0
six==1.16.0
typing_extensions==4.1.1
urllib3==1.26.9
varname==0.8.3
Werkzeug==2.1.1
wincertstore==0.2
